refactor structure and generation script
