+++ packages/ai-core/package.json
@@ -21,7 +21,7 @@
     "test": "NODE_OPTIONS=--experimental-vm-modules jest",
     "lint": "eslint . && prettier . --config ../../.prettierrc --ignore-path ../../.prettierignore -c",
     "lint:fix": "eslint . --fix && prettier . --config ../../.prettierrc --ignore-path ../../.prettierignore -w --log-level error",
-    "generate": "openapi-generator --generateESM  --overwrite --input ./src/spec/AI_CORE_API.yaml --outputDir ./src/client && pnpm lint:fix"
+    "generate": "openapi-generator --generateESM --clearOutputDir -i ./src/spec/AI_CORE_API.yaml -o ./src/client && pnpm lint:fix"
   },
   "dependencies": {
     "@sap-cloud-sdk/openapi": "^3.18.0"
